general changes
change data for socialicons.
and other minor changes
