+++ src/modules/Navigation/components/Burger/Burger.module.css
@@ -9,12 +9,12 @@
   transition: all 250ms cubic-bezier(0.5, 0, 0.5, 1);
 }
 
-.disNone {
+.displayNone {
   display: none;
 }
 
 @media (max-width: 961px) {
-  .disNone {
+  .displayNone {
     display: block;
   }
 }
