+++ src/modules/Navigation/components/Burger/components/BurgerCta/BurgerCta.js
@@ -1,6 +1,6 @@
 import React from "react"
 
-import { SocialIcons } from "modules/Navigation/components/common"
+import { SocialIcons } from "modules/common/components/SocialIcons"
 
 export function BurgerCta({ closeClickHandler }) {
     return <SocialIcons onClick={closeClickHandler} />
