+++ src/modules/Navigation/components/Navbar/Navbar.js
@@ -2,7 +2,7 @@ import React from "react"
 import styles from "./Navbar.module.css"
 
 import { NavigationList } from "../common"
-import { SocialIcons } from "../common"
+import { SocialIcons } from "modules/common/components/SocialIcons"
 
 export function Navbar() {
     return (
