+++ src/modules/Navigation/components/common/index.js
@@ -1,3 +1,2 @@
 export { NavigationLink } from "./NavigationLink"
 export { NavigationList } from "./NavigationList"
-export { SocialIcons } from "../../../common/components/SocialIcons"
