+++ src/modules/common/components/SocialIcons/data.js
@@ -1,0 +1,15 @@
+import { AiFillGithub, AiFillFacebook, AiFillLinkedin } from "react-icons/ai"
+
+export const socialsData = [
+    {
+        id: 1,
+        link: "https://www.facebook.com/profile.php?id=100085196048339",
+        icon: <AiFillFacebook />,
+    },
+    {
+        id: 2,
+        link: "https://www.linkedin.com/in/artur-pazyniuk-9a2bbb266/",
+        icon: <AiFillLinkedin />,
+    },
+    { id: 3, link: "https://github.com/Art1us", icon: <AiFillGithub /> },
+]
